refactor: move paths/openBrowser to src/infra in TypeScript
