+++ src/infra/openBrowser.ts
@@ -1,12 +1,23 @@
-const { spawn } = require('child_process');
+import { spawn } from 'child_process';
 
-function browserCommand(platform, url) {
+interface BrowserCommand {
+  cmd: string;
+  args: string[];
+}
+
+export function browserCommand(platform: NodeJS.Platform, url: string): BrowserCommand {
   if (platform === 'win32') return { cmd: 'cmd', args: ['/c', 'start', '', url] };
   if (platform === 'darwin') return { cmd: 'open', args: [url] };
   return { cmd: 'xdg-open', args: [url] };
 }
 
-function openBrowser(url, { platform = process.platform, env = process.env } = {}) {
+interface OpenBrowserOpts {
+  platform?: NodeJS.Platform;
+  env?: NodeJS.ProcessEnv;
+}
+
+export function openBrowser(url: string, opts: OpenBrowserOpts = {}): boolean {
+  const { platform = process.platform, env = process.env } = opts;
   if (env.NO_OPEN) return false;
   const { cmd, args } = browserCommand(platform, url);
   try {
@@ -18,5 +29,3 @@ function openBrowser(url, { platform = process.platform, env = process.env } = {
     return false;
   }
 }
-
-module.exports = { browserCommand, openBrowser };
